% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_factor}
\alias{fast_factor}
\title{creates a factor inspired by an idea of Kevin Ushey in the Rcpp gallery
http://gallery.rcpp.org/articles/fast-factor-generation/}
\usage{
fast_factor(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{charactervector with labels}
}
\description{
creates a factor inspired by an idea of Kevin Ushey in the Rcpp gallery
http://gallery.rcpp.org/articles/fast-factor-generation/
}
