% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{por}
\alias{por}
\title{Reads the binary SPSS file}
\usage{
por(filePath, debug)
}
\arguments{
\item{filePath}{The full systempath to the dta file you want to import.}
}
\description{
Reads the binary SPSS file
}
