% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readspss.R
\name{read.spss}
\alias{read.spss}
\title{read.spss}
\usage{
read.spss(x, ...)
}
\arguments{
\item{x}{file to import}

\item{...}{additional arguments passed to read.sav or read.por please see the
documentation for these functions.}
}
\description{
Function to read a SPSS (z)sav or por file into a data.frame(). This is just
a wrapper around read.sav and read.por for convenience.
}
\seealso{
\link{read.sav} and \link{read.por}
}
