% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readpor}
\alias{readpor}
\title{Reads the binary SPSS file}
\usage{
readpor(filePath, debug, encStr, override)
}
\arguments{
\item{filePath}{The full systempath to the dta file you want to import.}

\item{debug}{print debug information}

\item{encStr}{encoding string}

\item{override}{override bool}
}
\description{
Reads the binary SPSS file
}
