% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spss}
\alias{spss}
\title{Reads the binary SPSS file}
\usage{
spss(filePath, debug)
}
\arguments{
\item{filePath}{The full systempath to the dta file you want to import.}
}
\description{
Reads the binary SPSS file
}

