% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readpor.R
\name{read.por}
\alias{read.por}
\title{read.sav}
\usage{
read.por(file, convert.factors = TRUE, generate.factors = TRUE,
  encoding = TRUE, fromEncoding = NULL, use.missings = TRUE,
  debug = FALSE, override = FALSE, convert.dates = TRUE)
}
\description{
read.sav
}
