% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsav.R
\name{read.sav}
\alias{read.sav}
\title{read.sav}
\usage{
read.sav(file, convert.factors = TRUE, generate.factors = TRUE,
  encoding = NULL, fromEncoding = NULL, use.missings = TRUE,
  debug = FALSE, override = FALSE)
}
\arguments{
\item{file}{\emph{string} a sav-file to import. can be a file on a computer
or an url. in this case the file will be downloaded and read befor it is
used.}

\item{convert.factors}{\emph{logical} if true numeric or character variables
will be converted into a factor in R.}

\item{generate.factors}{\emph{logical} function to convert variables with
partial labels into factors. e.g. 1 - low and 5 - high are provided, labels
2, 3 and 4 will be created. especially useful in combination with
\code{use.missings=TRUE}.}

\item{encoding}{\emph{string} locale to convert to.}

\item{fromEncoding}{\emph{character.} encoding of the imported file. This
information is stored inside the sav-file, but is currently unused. Still
this option can be used to define the inital encoding by hand.}

\item{use.missings}{\emph{logical} should missing values be converted.
Defaults to TRUE.}

\item{debug}{\emph{logical} provides additional debug information. Most
likely not usefull to any user.}

\item{override}{\emph{logical}. The filename provided in \code{file} is
checked for the ending sav. If the fileending is different, nothing is read.
This option can be used to override this behavior.}
}
\value{
\code{readspss} returns a data.frame with additional objects
\describe{
\item{names}{colnames}
\item{row.names}{rownames}
\item{class}{data.frame}
\item{val.label}{value labels}
\item{datalabel}{datalabel}
\item{datestamp}{datestamp}
\item{timestamp}{timestamp}
\item{varmatrix}{a matrix with information how the data is stored}
\item{labnames}{list containing the information which variable uses which
label}
\item{missings}{a list containg information about the missing variables. if
\code{use.missings=TRUE} this Information will be used to generate missings.}
\item{vartype}{informations about a variable. is it numeric or string.}
}
}
\description{
Function to read a SPSS sav file into a data.frame().
}
\details{
SPSS files are widely available, though for R long time only foreign
and memisc provided functions to import sav-files. Lately haven joined.
This package is an approach to offer another alternative, to document the
sav-format and provide additional options to import the data.
}
\note{
information to decrypt the sav-format was provided by tda
\url{http://www.stat.ruhr-uni-bochum.de/tda.html} and pspp
 \url{http://www.gnu.org/software/pspp/}
}
\seealso{
\code{\link[foreign]{read.spss}}, \code{memisc} and
\code{\link[haven]{read_sav}}.
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}
}
