% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sav}
\alias{sav}
\title{Reads the binary SPSS file}
\usage{
sav(filePath, debug, encStr, ownEnc)
}
\arguments{
\item{filePath}{The full systempath to the dta file you want to import.}

\item{debug}{print debug information}

\item{encStr}{encoding string}

\item{ownEnc}{encoding provided by localeToCharset}
}
\description{
Reads the binary SPSS file
}
