% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writepor.R
\name{write.por}
\alias{write.por}
\title{write.por}
\usage{
write.por(dat, filepath, label)
}
\arguments{
\item{dat}{\emph{data.frame} a data.frame to export as por-file.}

\item{filepath}{\emph{string} full path where and how this file should be
stored}

\item{label}{\emph{character} vector of labels. must be of size `ncol(dat)`}
}
\value{
\code{readspss} returns nothing
}
\description{
Function to write an SPSS por file.
}
\details{
For now stores data.frames containing numerics only. Nothing else
 aside varnames and numerics are stored.
 Missing values in character cols (<NA>) are written as empty ("").
}
