% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writesav.R
\name{write.sav}
\alias{write.sav}
\title{read.sav}
\usage{
write.sav(dat, filepath, label)
}
\arguments{
\item{dat}{\emph{data.frame} a dat.aframe to strore as sav-file.}

\item{filepath}{\emph{string} full path where and how this file should be
stored}

\item{label}{\emph{character} vector of labels. must be of size `ncol(dat)`}
}
\value{
\code{readspss} returns nothing
}
\description{
Function to read a SPSS sav file into a data.frame().
}
\details{
For now stores data.frames containing numerics only. Nothing else
 aside varnames and numerics are stored.
 Missing values in character cols (<NA>) are written as empty ("").
}
