% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writesav}
\alias{writesav}
\title{writes the binary SPSS file}
\usage{
writesav(filePath, dat)
}
\arguments{
\item{filePath}{The full systempath to the dta file you want to import.}

\item{dat}{the data frame}
}
\description{
writes the binary SPSS file
}
